[syntax-errors] Make duplicate parameter names a semantic error
Summary
--

Moves `Parser::validate_parameters` to
`SemanticSyntaxChecker::duplicate_parameter_name`.

Test Plan
--

Existing tests, with `## Errors` replaced with `## Semantic Syntax Errors` and
an additional case for `lambda` expressions.
