+++ crates/ruff_linter/src/checkers/ast/mod.rs
@@ -615,7 +615,8 @@ impl SemanticSyntaxContext for Checker<'_> {
             | SemanticSyntaxErrorKind::DuplicateMatchKey(_)
             | SemanticSyntaxErrorKind::DuplicateMatchClassAttribute(_)
             | SemanticSyntaxErrorKind::InvalidStarExpression
-            | SemanticSyntaxErrorKind::AsyncComprehensionOutsideAsyncFunction(_) => {
+            | SemanticSyntaxErrorKind::AsyncComprehensionOutsideAsyncFunction(_)
+            | SemanticSyntaxErrorKind::DuplicateParameter(_) => {
                 if self.settings.preview.is_enabled() {
                     self.semantic_errors.borrow_mut().push(error);
                 }
